Separated main.js in multiple files
