--- src/linkHandler.js
@@ -1,8 +1,0 @@
-(function() {
-    document.addEventListener('click', (event) => {
-      if (event.target.tagName === 'A' && !event.target.href.startsWith('https://x.com')) {
-        event.preventDefault();
-        window.electron.openExternal(event.target.href);
-      }
-    }, true);
-  })();
